RecentsActivity Follow Recents Theme.
AOSP Follows what is an old style maybe

Froyo Gingerbread theme on Recents activity.

How this was discovered was playing with OTS on Recents shake to clean

Tasks and implemented OTS Dialog on Long clean on clear all button .

what came was the old gingerbread or froyo theme on the dialog.

with this recents follow the lollipop api and device default format.

Before :-

https://dl.pushbulletusercontent.com/HvILXtbtKpgfWFCg0qFV3msznK6OkdMP/download_20150819_204955.png

After :-

https://dl.pushbulletusercontent.com/6mm4TBDGTc99JWmtZ3ZDKJxxr8uoq6Or/Screenshot_2015-08-19-20-50-16.png

Change-Id: Ifd78ae964138e2c459249d3c853bf71be68c8126